refactor: connect database, test all and create Repository

+++ src/main/java/br/com/davi/spring_boot_first/dto/request/CriarProdutoRequest.java
@@ -3,13 +3,15 @@ package br.com.davi.spring_boot_first.dto.request;
 import lombok.Getter;
 import lombok.Setter;
 
-    @Getter
+import java.math.BigDecimal;
+
+@Getter
     @Setter
 
 
 // o que o usuário envia (tipo Schema)
     public class CriarProdutoRequest {
         private String name;
-        private double preco;
-        private int quantidade;
+        private BigDecimal preco;
+        private Integer quantidade;
     }
