+++ src/main/java/br/com/davi/spring_boot_first/dto/request/EditarProdutoRequest.java
@@ -7,6 +7,6 @@ import lombok.*;
 
 
 public class EditarProdutoRequest {
-        private int id;
-        private int quantidade;
+        private Long id;
+        private Integer quantidade;
     }
