+++ src/main/java/br/com/davi/spring_boot_first/dto/response/EditarProdutoResponse.java
@@ -9,7 +9,7 @@ import lombok.*;
 
 
 public class EditarProdutoResponse {
-    private int id;
+    private Long id;
     private String nome;
-    private int quantidade;
+    private Integer quantidade;
 }
